+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- BLIDE-271 Improve interface method diagnostics
 - Improve project indexing
 - Add API for querying the project index
 - BLIDE-224 Add hover support for variables, functions, classes, and interfaces